fix(menu/pause): handle box looping when Dialog buttons pressed
For example, when the 'Extra' button is pressed, the menu box will exit
with code 3, which will cause the box to exit when it's supposed to
continue loop.

The box should stop looping only when 'Cancel' button or ESC button is
pressed, or internal (renderer) errors occurred.

Both menu and pause boxes were affected, as they have a custom looper.

Signed-off-by: iusmac <iusico.maxim@libero.it>