fix(core): create a FIFO per user for capturing stderr from the renderer
Since we re-create the FIFO file for capturing stderr from the renderer
every time the app is started, there's a chance that we might be unable
to do that if the file was created for a different user, such as root,
if user used sudo to start the app.

This patchfix appends UID to file name to prevent potential clashes and
permission issues.

Signed-off-by: iusmac <iusico.maxim@libero.it>