fix(program): prevent long text from wrapping (Whiptail)
The Whiptail's info box used to display the command's output, wraps text
that exceeds the box width, but Dialog's program box limits the lines to
the box width.

Wrapping prevents the user from fully viewing the command's output, as
wrapped lines eat space in the scrolling area, which means the newer
text will be "out of sight".

As a workaround, we'll limit all lines to the box width ourselves.

Run demo/program.sh to reproduce.

┌───────┤ Before ├────────┐ ┌────────┤ After ├────────┐
│                         │ │                         │
│ Display the output of a │ │ Display the output of a │
│ command using a pipe    │ │ command using a pipe    │
│                         │ │                         │
│ Example long scrolling  │ │ Example long scrolling  │
│ output (line 19)        │ │ Example long scrolling  │
│ Example long scrolling  │ │ Done                    │
│                         │ │                         │
│         <Ok>            │ │         <Ok>            │
│                         │ │                         │
└─────────────────────────┘ └─────────────────────────┘

Signed-off-by: iusmac <iusico.maxim@libero.it>