fix: (actually) support `listHeight=auto` & `menuHeight=auto` options
Both Dialog and Whiptail support auto-sizing of list and menu boxes to
fit the contents (entries) when passing 0 as third argument. Example
for a menu box:

    --menu text height width menu-height [ tag item ] ...

For better UX, we had planned to support listHeight='auto' and
menuHeight='auto', which simply translates 'auto' to 0. This was
documented but not implemented for both list box and menu box.

Signed-off-by: iusmac <iusico.maxim@libero.it>