doc(readme): replace the demo video with a HQ GIF image
GitHub's video embedding feature follows the lazy loading mechanism,
which causes to change the page layout in the wrong moment, which seems
to confuse the browser when it should perform a jump down to a #Section
of the README via anchor links.

We'll use a GIF instead of a MP4 as a reasonable workaround.

This also visually improves the demo, as we can use the yuv444p pixel
format (24 bits/px) to process the frames, while the MP4 is forced to
follow the yuv420p pixel format (12 bits/px) in order to play in the
browser (especially the mobile ones).

The demo/images/gendemo.sh script was written to generate a HQ GIF
images that is saved in demo/images/demo.gif.

Signed-off-by: iusmac <iusico.maxim@libero.it>