fix(box): correctly process the per-result callbacks
The previous conditional check (i < ${#l_result[@]) is wrong, as the
index (i) increments as we go, but the number of remaining results may
not always be lined up with the index variable if there are more results
than callbacks, or, if the callbacks array is fragmented compared to
results, e.g., a checklist in which 2/3 elements have been selected:

      results[]  |   callbacks[]
    ------------ | ----------------
    [0]='Test 1' | ['Test 1']='cb1()'
    [1]='Test 2' | ['Test 2']='cb2()'
    [2]='Test 3' |

The previous conditional check may also prevent from processing all of
the results, as the index may be equal to the remaining results.

Instead, we should loop until we have callbacks and stop as soon as
there are no more remaining results.

Signed-off-by: iusmac <iusico.maxim@libero.it>