fix(program): address memory leak
The 'local l_text_old="$l_text"' WILL NOT overwrite the previous result
when performing redefine+reassign like we're doing when looping over the
new lines. As a result, we only accumulate lines that are outside of the
scrollable area anyway, so memory leak occurs. Furthermore, it also adds
work to tail command.

Signed-off-by: iusmac <iusico.maxim@libero.it>