fix(range): add missing support for `+=` operator in the `text` option
The range box hijacks the 'text=...' option instead of propagating it
"as-is" to the parent. Only the 'text+=...' option will effectively
propagate to the parent, but the value will be overridden by the
'text=...' option.

The support for concatenation ('+=') operator in the 'text' option was
added in ba800eb9bab7b3125a27e12e826431ea44739b64.

Signed-off-by: iusmac <iusico.maxim@libero.it>